requirements.txt
torch
numpy
scipy
nltk
scikit-learn
pyyaml
tqdm
